Consolidate customer getter/setter methods
