+++ src/constants.tsx
@@ -1,8 +1,8 @@
 /**
  * Screen names as used in navigation. Note that for typing purposes these values must correspond exactly to
- * the key names in the `RootStackParamList`. If these names are ever updated, the param list should be 
+ * the key names in the `RootStackParamList`. If these names are ever updated, the param list should be
  * updated as well to match.
- * 
+ *
  * @see {@link RootStackParamList}
  */
 export enum Screen {
@@ -12,8 +12,4 @@ export enum Screen {
   ViewCustomer = "ViewCustomer",
   EditCustomer = "EditCustomer",
   AddCustomer = "AddCustomer",
-}
-
-export enum AsyncStorageKey {
-  Customer = "CUSTOMER",
 }
