+++ src/store/sagas/customers/deleteCustomerSaga.tsx
@@ -2,8 +2,10 @@ import { put, takeLatest, delay } from "redux-saga/effects";
 import * as actions from "../../reducers/customersReducer";
 import { Customer } from "../../reducers/customersReducer";
 import { PayloadAction } from "@reduxjs/toolkit";
-import AsyncStorage from "@react-native-async-storage/async-storage";
-import { AsyncStorageKey } from "../../../constants";
+import {
+  getCustomersFromStorage,
+  setCustomersToStorage,
+} from "../../utilities/async_storage";
 
 export function* watchDeleteCustomer() {
   yield takeLatest(actions.deleteCustomer.toString(), takeDeleteCustomer);
@@ -28,16 +30,3 @@ export function* takeDeleteCustomer(action: PayloadAction<number>) {
     yield put(actions.deleteCustomerError(error.toString()));
   }
 }
-
-const getCustomersFromStorage = async () => {
-  const customerData = await AsyncStorage.getItem(AsyncStorageKey.Customer);
-  return customerData ? JSON.parse(customerData) : [];
-};
-
-const setCustomersToStorage = async (customers: Customer[]) => {
-  const serializedCustomers = JSON.stringify(customers);
-  return await AsyncStorage.setItem(
-    AsyncStorageKey.Customer,
-    serializedCustomers
-  );
-};
