+++ src/store/sagas/customers/saveCustomerSaga.tsx
@@ -2,8 +2,10 @@ import { put, takeLatest, delay } from "redux-saga/effects";
 import * as actions from "../../reducers/customersReducer";
 import { Customer } from "../../reducers/customersReducer";
 import { PayloadAction } from "@reduxjs/toolkit";
-import AsyncStorage from "@react-native-async-storage/async-storage";
-import { AsyncStorageKey } from "../../../constants";
+import {
+  getCustomersFromStorage,
+  setCustomersToStorage,
+} from "../../utilities/async_storage";
 
 export function* watchSaveCustomer() {
   yield takeLatest(actions.saveCustomer.toString(), takeSaveCustomer);
@@ -59,17 +61,4 @@ const editCustomer = (
   }
 
   return existingCustomers;
-};
-
-const getCustomersFromStorage = async () => {
-  const customerData = await AsyncStorage.getItem(AsyncStorageKey.Customer);
-  return customerData ? JSON.parse(customerData) : [];
-};
-
-const setCustomersToStorage = async (customers: Customer[]) => {
-  const serializedCustomers = JSON.stringify(customers);
-  return await AsyncStorage.setItem(
-    AsyncStorageKey.Customer,
-    serializedCustomers
-  );
 };
