+++ src/store/sagas/customers/syncCustomersSaga.tsx
@@ -1,8 +1,7 @@
 import { put, takeLatest, delay } from "redux-saga/effects";
 import * as actions from "../../reducers/customersReducer";
 import { Customer } from "../../reducers/customersReducer";
-import AsyncStorage from "@react-native-async-storage/async-storage";
-import { AsyncStorageKey } from "../../../constants";
+import { getCustomersFromStorage } from "../../utilities/async_storage";
 
 export function* watchSyncCustomers() {
   yield takeLatest(actions.syncCustomers.toString(), takeSyncCustomers);
@@ -17,8 +16,3 @@ export function* takeSyncCustomers() {
     yield put(actions.syncCustomersResult(error.toString()));
   }
 }
-
-const getCustomersFromStorage = async () => {
-  const customerData = await AsyncStorage.getItem(AsyncStorageKey.Customer);
-  return customerData ? JSON.parse(customerData) : [];
-};
